requirements.txt
asgiref==3.2.3
Django==1.6.4
djangorestframework==3.0.0
pytz==2020.4
sqlparse==0.3.1
